ocrd_monitor/flask/dev-requirements.txt
pytest==7.1.3
pytest-playwright==0.3.0